cpu/fe310: reorganize files and includes
